refactor: update mui-fine-foods notistack with lazy initialization
